move the media session handling from the PlaybackServiceMediaPlayer and into the PlaybackService
